Integrate DSDL selection in setup window
Iface configuration and DSDL selection are both done in
the same setup window. This window is also bound to
disappear in the near future. Configuration will be
possible through the main app window instead.
